--- packages/browser-wallet/src/popup/popupX/pages/TransactionDetails/i18n/en.ts
@@ -1,11 +1,0 @@
-const t = {
-    backTitle: 'to Activity',
-    title: 'Transaction details',
-    tHash: ' Transaction hash',
-    bHash: 'Block hash',
-    from: 'From address',
-    to: 'To address',
-    events: 'Events',
-};
-
-export default t;
