+++ packages/browser-wallet/src/popup/popupX/pages/TransactionLog/TransactionDetails/index.ts
