+++ packages/browser-wallet/src/popup/popupX/styles/_elements.scss
@@ -37,12 +37,12 @@
 @import '../pages/SeedPhrase/SeedPhrase';
 @import '../pages/ChangePasscode/ChangePasscode';
 @import '../pages/TransactionLog/TransactionLog';
+@import '../pages/TransactionLog/TransactionDetails/TransactionDetails';
 @import '../pages/Web3Id/Web3Id';
 @import '../pages/MainPage/MainPage';
 @import '../pages/TokenDetails/TokenDetails';
 @import '../pages/ManageTokens/ManageTokens';
 @import '../pages/SendFunds/SendFunds';
-@import '../pages/TransactionDetails/TransactionDetails';
 @import '../pages/Nft/Nft';
 @import '../pages/EarningRewards/EarningRewards';
 @import '../pages/EarningRewards/AccountCooldowns/AccountCooldowns';
